+++ integrationtest/kube/collect_kube_test.go
@@ -590,6 +590,11 @@ dremio-jfr-time-seconds: 10
 	t.Logf("checking file %v", filepath.Join(hcDir, "kubernetes", "nodes", "describe-nodes.txt"))
 	tests.AssertFileHasExpectedLines(t, []string{"Name:", "Roles:", "Labels:"}, filepath.Join(hcDir, "kubernetes", "nodes", "describe-nodes.txt"))
 
+	// check cluster describe pod files
+	tests.AssertFileHasContent(t, filepath.Join(hcDir, "kubernetes", "pods", "describe-pods.txt"))
+	t.Logf("checking file %v", filepath.Join(hcDir, "kubernetes", "[pds]", "describe-pods.txt"))
+	tests.AssertFileHasExpectedLines(t, []string{"Labels:", "Annotations:", "Containers:"}, filepath.Join(hcDir, "kubernetes", "pods", "describe-pods.txt"))
+
 	//kvstore report
 	tests.AssertFileHasContent(t, filepath.Join(hcDir, "kvstore", "dremio-master-0", "kvstore-report.zip"))